feat: Use .env file for environment variables
This commit introduces the use of a .env file to manage environment variables.
The django-environ library has been added to the project to handle the loading of these variables.
The ALLOWED_HOSTS setting is now loaded from the .env file.
The .env file has been added to .gitignore to prevent it from being committed to the repository.